Refactor: Extract demo sections to separate files
- Create data/business-demo-sections.ts for business planning demo data
- Create data/tech-demo-sections.ts for technical specification demo data
- Update page.tsx and code/page.tsx to import from new files
- Reduce context size by moving large demo objects out of components
